refactor profile header
