+++ .env
@@ -1,4 +1,3 @@
-PUBLIC_BASE_URL=https://server-alpha.tech
-# PUBLIC_BASE_URL=http://localhost:3000
-# PUBLIC_BASE_URL=https://server-beta.de
-
+#PUBLIC_BASE_URL=https://server-alpha.tech
+PUBLIC_BASE_URL=http://localhost:3000
+#PUBLIC_BASE_URL=https://server-beta.de
